requirements.txt
flask
psycopg2-binary
Flask-SQLAlchemy
python-dotenv==0.19.1